requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
aiofiles==23.2.1
